frontend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
pydantic==2.11.9
numpy==2.3.3
pandas==2.3.2
pillow==11.3.0
joblib==1.5.2
scikit-learn==1.7.2
lightgbm==4.6.0
torch==2.6.0
transformers==4.39.3
huggingface-hub==0.22.2
tokenizers==0.15.2
safetensors==0.5.3
